Move `StoryDetailsUIComposer` into its own file
